+++ CHANGELOG.md
@@ -18,5 +18,7 @@
 * **breaking**: Put internal system methods in to Ext traits, expose underlying types with crate specific Ext traits.
 * input: MacOS scancode normalization
 * riddle: Relax bounds on error type param for `RiddleLib::run_with_err` to allow use with `anyhow::Error`.
+* **breaking**: Add clippy to the CI, and fix a variety of issues. Visible change is `new` methods which return a handle
+    type have been renamed `new_shared`.
 
 ## 0.1.0 - Initial Release